refactor: 一気にドカンとリファクタリング
